Reorganize protocol stuff based on current specs
