+++ packages/portalnetwork/src/subprotocols/canonicalIndices/index.ts
@@ -1,0 +1,1 @@
+export * from './canonicalIndices'
