--- packages/portalnetwork/src/subprotocols/headerGossip/index.ts
@@ -1,3 +1,0 @@
-export * from './types'
-export * from './util'
-export * from './headerAccumulator'
